#0000 Platform: Optimization / enhancement of function compact

+++ platform/Worker.cs
@@ -322,7 +322,8 @@ namespace Platform {
                             }
 
                             string recycleDirectory = Path.Combine(workerTask.Drive, "$RECYCLE.BIN");
-                            Directory.Delete(recycleDirectory, true);
+                            if (Directory.Exists(recycleDirectory))
+                                Directory.Delete(recycleDirectory, true);
 
                             string startupExitFile = Path.Combine(workerTask.Drive, "Startup.exit");
                             if (File.Exists(startupExitFile))
